+++ ksp/generator/src/main/kotlin/KSClassDeclarationCompanion.kt
@@ -1,0 +1,8 @@
+package dev.inmo.micro_ksp.generator
+
+import com.google.devtools.ksp.symbol.KSClassDeclaration
+
+val KSClassDeclaration.companion
+    get() = declarations.firstNotNullOfOrNull {
+        (it as? KSClassDeclaration)?.takeIf { it.isCompanionObject }
+    }